Refactor tool discovery logic and associated unit tests
